migrator/src/requirements.txt
pymongo==4.1.0
python-dotenv==0.15.0
psycopg2-binary==2.9.3
requests==2.25.1